chore: domain 패키지 구분
