+++ src/main/java/domain/GameManager.java
@@ -1,5 +1,13 @@
 package domain;
 
+import domain.card.CardProvider;
+import domain.card.Cards;
+import domain.participant.Dealer;
+import domain.participant.Participant;
+import domain.participant.ParticipantName;
+import domain.participant.Participants;
+import domain.participant.Player;
+
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
