+++ src/main/java/domain/card/Suit.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.Arrays;
 import java.util.List;
