+++ src/main/java/domain/participant/ParticipantName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import java.util.Objects;
