+++ src/test/java/domain/card/CardsTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
